backend/requirements.txt
fastapi
uvicorn
pdfplumber
google-genai
python-multipart
python-dotenv
